refactor: split some of the types into their own files
