+++ src/profile/models/mod.rs
@@ -1,1 +1,4 @@
+pub mod focus_school;
+pub mod load_out_preset;
+pub mod platform;
 pub(crate) mod profile;
